% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{copy_myriad_fonts}
\alias{copy_myriad_fonts}
\title{Copy Myriad Pro and Socviz fonts from ~/Library/Fonts to the myriad package}
\usage{
copy_myriad_fonts()
}
\value{
Puts the fonts in the right place for use with this package
}
\description{
Not a portable solution. Only works on MacOS. Assumes you have the fonts
installed in ~/Library/Fonts
}
\examples{
\dontrun{
}
}
