% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{import_myriad_condensed}
\alias{import_myriad_condensed}
\title{Import Myriad Condensed only}
\usage{
import_myriad_condensed()

import_myriad_condensed()
}
\value{
Makes "Myriad Condensed" available for use

Makes "Socviz Condensed" available for use
}
\description{
Import Myriad Condensed only

Import Socviz Condensed
}
\examples{
\dontrun{
}
\dontrun{
}
}
