% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{import_myriad_condensed_pdf}
\alias{import_myriad_condensed_pdf}
\title{Import Myriad Condensed for used in PDFs}
\usage{
import_myriad_condensed_pdf()
}
\value{
Makes "Myriad Condensed PDF" available for use in PDFs only
}
\description{
Import Myriad Condensed for used in PDFs
}
\examples{
\dontrun{
}
}
