% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{import_myriad_pdf}
\alias{import_myriad_pdf}
\title{Import Myriad SemiCondensed for use in PDFs}
\usage{
import_myriad_pdf()
}
\value{
Makes "Myriad Semi PDF" available for use in PDFs only
}
\description{
Import Myriad SemiCondensed for use in PDFs
}
\examples{
\dontrun{
}
}
