% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{import_socviz_condensed}
\alias{import_socviz_condensed}
\title{Import Socviz Condensed}
\usage{
import_socviz_condensed()
}
\value{
Makes "Socviz Condensed" available for use
}
\description{
Import Socviz Condensed
}
\examples{
\dontrun{
}
}
