% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{import_socviz_semi}
\alias{import_socviz_semi}
\title{Import Socviz Semicondensed}
\usage{
import_socviz_semi()
}
\value{
Makes "Socviz SemiCondensed" available for use
}
\description{
Import Socviz Semicondensed
}
\examples{
\dontrun{
}
}
