% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myriad.r
\docType{data}
\name{font_mc}
\alias{font_mc}
\title{Myriad Pro Condensed font name R variable aliases}
\format{length 1 character vector}
\usage{
font_mc
}
\description{
\code{font_myriad} == "\code{Myriad Pro Condensed}"
}
\keyword{datasets}
