% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myriad.r
\docType{data}
\name{font_myriad_light}
\alias{font_myriad_light}
\title{Myriad Pro Light Condensed font name R variable aliases}
\format{
An object of class \code{character} of length 1.
}
\usage{
font_myriad_light
}
\description{
\code{font_fc_light} == "\verb{Myriad Pro Light Condensed}"
}
\keyword{datasets}
