% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{pdf_ggsave}
\alias{pdf_ggsave}
\alias{ggsave_pdf}
\title{Hack ggsave for use with showtext when making PDFs}
\usage{
pdf_ggsave(
  filename,
  plot = last_plot(),
  device = "pdf",
  path = NULL,
  scale = 1,
  width = NA,
  height = NA,
  units = c("in", "cm", "mm", "px"),
  dpi = 300,
  limitsize = TRUE,
  bg = NULL,
  ...
)

ggsave_pdf(
  filename,
  plot = last_plot(),
  device = "pdf",
  path = NULL,
  scale = 1,
  width = NA,
  height = NA,
  units = c("in", "cm", "mm", "px"),
  dpi = 300,
  limitsize = TRUE,
  bg = NULL,
  ...
)
}
\arguments{
\item{filename}{ggsave}

\item{plot}{ggsave}

\item{device}{ggsave}

\item{path}{ggsave}

\item{scale}{ggsave}

\item{width}{ggsave}

\item{height}{ggsave}

\item{units}{ggsave}

\item{dpi}{ggsave}

\item{limitsize}{ggsave}

\item{bg}{ggsave}

\item{...}{ggsave}
}
\value{
A showtexted plot
}
\description{
This is gross
}
\examples{
\dontrun{
}
\dontrun{
ggsave_pdf()
}
}
