% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myriad.r
\name{theme_myriad_map}
\alias{theme_myriad_map}
\title{theme_myriad_map}
\usage{
theme_myriad_map()
}
\description{
Myriad map theme
}
\details{
You should [import_myriad_semi]() first and also install the fonts on your
system before trying to use this theme.
}
\examples{
\dontrun{
}
}
