% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myriad.r
\name{theme_myriad_nymap}
\alias{theme_myriad_nymap}
\title{Myriad New York City map theme}
\usage{
theme_myriad_nymap()
}
\description{
Map theme suitable for NYC maps. As [theme_myriad_map()], but with the
legend placed inside the plot area.
}
\examples{
\dontrun{
}
}
