% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myriad.r
\name{theme_myriad_nymap_pdf}
\alias{theme_myriad_nymap_pdf}
\title{Myriad New York City map theme PDF}
\usage{
theme_myriad_nymap_pdf()
}
\description{
Map theme suitable for NYC maps.
You should [import_myriad_semi]() first and also install the fonts on your
system before trying to use this theme.
}
\examples{
\dontrun{
}
}
