% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myriad.r
\name{theme_myriad_semi_pdf}
\alias{theme_myriad_semi_pdf}
\title{PDF/showtext version of theme_myriad_semi()}
\usage{
theme_myriad_semi_pdf(
  base_family = "Myriad Pro PDF",
  title_family = "Myriad Pro PDF"
)
}
\arguments{
\item{base_family}{Base font}

\item{title_family}{Title font}
}
\value{
ggplot2 theme
}
\description{
PDF/showtext version of theme_myriad_semi()
}
\examples{
\dontrun{

}
}
