% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{import_myriad}
\alias{import_myriad}
\title{Import all Myriad Pro font faces and styles for use in graphs
as "Myriad Pro".}
\usage{
import_myriad(font_family = "Myriad Pro", silent = TRUE)
}
\arguments{
\item{font_family}{The font family you want to import}

\item{silent}{Report on what's been imported or not}
}
\description{
This will work for devices that use `systemfonts` to process fonts.
For PDFs, use showtext::showtext_auto() beforehand
}
\note{
Modified from https://yjunechoe.github.io/posts/2021-06-24-setting-up-and-debugging-custom-fonts/.
}
