% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myriad.r
\name{import_myriad}
\alias{import_myriad}
\title{Import Myriad font for use in charts}
\usage{
import_myriad()
}
\description{
Import Myriad font for use in charts
}
\note{
This will take care of ensuring PDF/PostScript usage. The location of the
  font directory is displayed after the base import is complete. It is highly
  recommended that you install them on your system the same way you would any
  other font you wish to use in other programs.
}
