% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{import_myriad}
\alias{import_myriad}
\alias{import_myriad_semi}
\title{Import all Myriad Pro font faces and styles for use in graphs
This will work for devices that use systemfonts to process fonts
But it won't work for PDFs, including ggsave() where the output is a PDF}
\usage{
import_myriad(family = "Myriad Pro", silent = TRUE)

import_myriad_semi(family = "Myriad Pro", silent = TRUE)
}
\arguments{
\item{family}{The font family you want to import}

\item{silent}{Report on what's been imported or not}
}
\description{
Import all Myriad Pro font faces and styles for use in graphs
This will work for devices that use systemfonts to process fonts
But it won't work for PDFs, including ggsave() where the output is a PDF
}
\note{
From https://yjunechoe.github.io/posts/2021-06-24-setting-up-and-debugging-custom-fonts/.
}
\examples{
\dontrun{
import_myriad_semi()
}
}
