% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{import_myriad_semi}
\alias{import_myriad_semi}
\title{Import Myriad SemiCondensed only}
\usage{
import_myriad_semi()
}
\value{
Makes "Myriad SemiCondensed" available for use
}
\description{
Import Myriad SemiCondensed only
}
\examples{
\dontrun{
}
}
