% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myriad.r
\name{theme_myriad_semi}
\alias{theme_myriad_semi}
\title{theme_myriad_semi}
\usage{
theme_myriad_semi(
  base_size = 12,
  base_family = "Myriad Pro Condensed",
  title_family = "Myriad Pro Semibold SemiCondensed",
  base_line_size = base_size/24,
  base_rect_size = base_size/24
)
}
\arguments{
\item{base_family, title_family, base_size, base_line_size, base_rect_size}{base font family and size}
}
\description{
A [ggplot2] theme using semibold variants of Adobe Myriad Pro
}
\details{
You should [import_myriad_semi]() first and also install the fonts on your
system before trying to use this theme.
}
\examples{
\dontrun{
}
}
\author{
Kieran Healy
}
