% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myriad.r
\name{theme_socviz_semi}
\alias{theme_socviz_semi}
\title{theme_socviz_semi}
\usage{
theme_socviz_semi(
  base_size = 12,
  base_family = "Socviz SemiCondensed",
  title_family = "Socviz SemiCondensed",
  base_line_size = base_size/24,
  base_rect_size = base_size/24
)
}
\arguments{
\item{base_family, title_family, base_size, base_line_size, base_rect_size}{base font family and size}
}
\description{
A [ggplot2] theme using semibold variants of Adobe Myriad Pro
}
\examples{
\dontrun{
}
}
\author{
Kieran Healy
}
